requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.5.6
black==21.5b1
certifi==2020.12.5
click==8.0.0
greenlet==1.1.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
pathspec==0.8.1
pylint==2.8.2
pyls==0.1.6
pynvim==0.4.3
python-telegram-bot==13.5
pytz==2021.1
regex==2021.4.4
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
wrapt==1.12.1
